perf(web/mini-apps/memory-game): improve UI rendering
