+++ apps/web/app/(feat)/mini-apps/memory-game/assets/components/memory-card-container.tsx
@@ -1,101 +1,17 @@
-"use client";
-
-import React, { useEffect, useState } from "react";
+import React from "react";
 import MemoryCardItem from "./memory-card-item";
 import { DifficultyLevel } from "./memory-level";
 import Button from "@/components/niwi-ui/button/button";
+import CongratulationsBox from "./congratulation-box";
+import { useMemoryGame } from "../hooks/useMemoryGame";
 
-type GameValue = number;
-
-type Card = {
-  id: number;
-  value: number;
-  color: string;
-  isMatched: boolean;
-  isFlipped: boolean;
-};
-
-const shuffleArray = (array: any) => {
-  for (let i = array.length - 1; i > 0; i--) {
-    const j = Math.floor(Math.random() * (i + 1));
-    [array[i], array[j]] = [array[j], array[i]];
-  }
-  return array;
-};
-
-const generatePairsWithColors = (numPairs: number, level: DifficultyLevel) => {
-  const values: Card[] = [];
-  const difficultyColors: Record<DifficultyLevel, string[]> = {
-    EASY: ["#FF5733", "#33FF57", "#3357FF", "#F3FF33", "#FF33F3", "#33F3FF"],
-    MEDIUM: [
-      "#FF5733",
-      "#33FF57",
-      "#3357FF",
-      "#F3FF33",
-      "#FF33F3",
-      "#33F3FF",
-      "#FF8C00",
-      "#8A2BE2",
-      "#7FFF00",
-      "#D2691E",
-      "#DC143C",
-      "#FF4500",
-    ],
-    HARD: [
-      "#FF5733",
-      "#33FF57",
-      "#3357FF",
-      "#F3FF33",
-      "#FF33F3",
-      "#33F3FF",
-      "#FF8C00",
-      "#8A2BE2",
-      "#7FFF00",
-      "#D2691E",
-      "#DC143C",
-      "#FF4500",
-      "#FFD700",
-      "#ADFF2F",
-      "#FF1493",
-      "#C71585",
-      "#FF69B4",
-      "#FF6347",
-      "#FF4500",
-      "#2E8B57",
-      "#4682B4",
-      "#B0C4DE",
-    ],
-  };
-
-  let colors = difficultyColors[`${level}`];
-
-  for (let i = 1; i <= numPairs; i++) {
-    const color = colors[i % colors.length]!;
-    values.push({
-      id: Math.ceil(Math.random() * 100000) + new Date().getTime() + 1,
-      value: i,
-      color,
-      isFlipped: false,
-      isMatched: false,
-    });
-    values.push({
-      id: Math.ceil(Math.random() * 100000) + new Date().getTime() + 2,
-      value: i,
-      color,
-      isFlipped: false,
-      isMatched: false,
-    });
-  }
-
-  return shuffleArray(values);
-};
-
-const MemoryCardContainer = ({ level }: { level: DifficultyLevel }) => {
-  const [cards, setCards] = useState<Card[]>([]);
-  const [selectedCards, setSelectedCards] = useState<Card[]>([]);
-  const [matchedCards, setMatchedCards] = useState<Set<number>>(new Set());
-  const [gameOver, setGameOver] = useState<boolean>(false);
-
+const MemoryCardContainer = ({
+  level,
+  setLevel,
+}: {
+  level: DifficultyLevel;
+  setLevel: (value: DifficultyLevel | null) => void;
+}) => {
   let numPairs: number;
 
   switch (level) {
@@ -113,60 +29,14 @@ const MemoryCardContainer = ({ level }: { level: DifficultyLevel }) => {
       break;
   }
 
-  useEffect(() => {
-    const pairsArray = generatePairsWithColors(numPairs, level);
-    setCards(pairsArray);
-  }, [level]);
+  const { cards, gameOver, handleCardClick } = useMemoryGame(level, numPairs);
 
-  useEffect(() => {
-    if (selectedCards.length === 2) {
-      const [firstCard, secondCard] = selectedCards;
-
-      if (firstCard?.value === secondCard?.value) {
-        setCards((prevCards) =>
-          prevCards.map((card) =>
-            card.value === firstCard?.value
-              ? { ...card, matched: true, isFlipped: true }
-              : card
-          )
-        );
-        setMatchedCards((prev) => new Set(prev).add(firstCard!.value));
-      } else {
-        setTimeout(() => {
-          setCards((prevCards) =>
-            prevCards.map((card) =>
-              card.isFlipped! && !card.isMatched
-                ? { ...card, isFlipped: false }
-                : card
-            )
-          );
-        }, 1000);
-      }
-      setSelectedCards([]);
-    }
-  }, [selectedCards]);
-
-  useEffect(() => {
-    if (matchedCards.size == cards.length) {
-      setGameOver(true);
-    }
-  }, [selectedCards]);
-
-  const handleCardClick = (id: number) => {
-    const card = cards.find((c) => c.id === id);
-    if (card && !card.isFlipped && selectedCards.length < 2) {
-      setCards((prevCards) =>
-        prevCards.map((c) => (c.id === id ? { ...c, isFlipped: true } : c))
-      );
-      setSelectedCards((prev) => [...prev, card]);
-    }
-  };
-
-  return (
-    <div className="w-full px-10 grid grid-cols-6 gap-10">
-      {cards.map((item, index) => {
-        console.log(item);
-        return (
+  return gameOver ? (
+    <CongratulationsBox setLevel={setLevel} />
+  ) : (
+    <>
+      <div className="w-full px-10 grid grid-cols-6 gap-10 my-10">
+        {cards.map((item, index) => (
           <MemoryCardItem
             key={index}
             num={item.value}
@@ -174,10 +44,17 @@ const MemoryCardContainer = ({ level }: { level: DifficultyLevel }) => {
             isFlipped={item.isFlipped}
             onClick={() => handleCardClick(item.id)}
           />
-        );
-      })}
-      <Button variant={"niwi"}>Change Level</Button>
-    </div>
+        ))}
+      </div>
+      <Button
+        variant={"niwi"}
+        onClick={() => {
+          setLevel(null);
+        }}
+      >
+        Change Level
+      </Button>
+    </>
   );
 };
