+++ apps/web/app/(feat)/mini-apps/memory-game/assets/components/memory-level.tsx
@@ -6,9 +6,9 @@ import MemoryCardContainer from "./memory-card-container";
 export type DifficultyLevel = "EASY" | "MEDIUM" | "HARD";
 
 const MemoryLevel: React.FC = () => {
-  const [level, setLevel] = useState<DifficultyLevel | undefined>(undefined);
+  const [level, setLevel] = useState<DifficultyLevel | null>(null);
 
-  const handleLevelSelect = (difficulty: DifficultyLevel) => {
+  const handleLevelSelect = (difficulty: DifficultyLevel | null) => {
     setLevel(difficulty);
   };
 
@@ -16,7 +16,7 @@ const MemoryLevel: React.FC = () => {
     <section>
       {!level ? (
         <div className="grid gap-5">
-          <h1 className="niwi-logo-text mx-auto">Please Select Difficulty</h1>
+          <h1 className="niwi-logo-text mx-auto">Welcome To Memory Game </h1>
           <div className="w-[20rem] px-10 grid place-items-center gap-3 mx-auto">
             <Button
               onClick={() => handleLevelSelect("EASY")}
@@ -40,9 +40,18 @@ const MemoryLevel: React.FC = () => {
               Hard
             </Button>
           </div>
+          <p className="text-sm mx-auto">
+            <span className="text-yellow-500/60 mx-2">Warning-</span>
+            <span className="text-gray-500">
+              Control your mind and breathe first
+            </span>
+          </p>
         </div>
       ) : (
-        <MemoryCardContainer level={level} />
+        <MemoryCardContainer
+          level={level}
+          setLevel={(level: DifficultyLevel | null) => handleLevelSelect(level)}
+        />
       )}
     </section>
   );
